requirements.txt
aniso8601==7.0.0
appdirs==1.4.4
argon2-cffi==21.1.0
asgiref @ file:///tmp/build/80754af9/asgiref_1625643473416/work
astroid @ file:///C:/ci/astroid_1613501150380/work
attrs @ file:///tmp/build/80754af9/attrs_1620827162558/work
autopep8 @ file:///tmp/build/80754af9/autopep8_1620866417880/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
bcrypt @ file:///C:/ci/bcrypt_1607022693089/work
black==19.10b0
blinker==1.4
brotlipy==0.7.0
certifi==2021.5.30
cffi @ file:///C:/ci/cffi_1625831756778/work
chardet @ file:///D:/bld/chardet_1610093673558/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click @ file:///tmp/build/80754af9/click_1621604852318/work
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
coreapi @ file:///home/conda/feedstock_root/build_artifacts/coreapi_1596221916599/work
coreschema @ file:///home/conda/feedstock_root/build_artifacts/coreschema_1596220533330/work
coverage @ file:///C:/ci/coverage_1614615100481/work
cryptography @ file:///C:/ci/cryptography_1616769504165/work
decorator @ file:///tmp/build/80754af9/decorator_1621259047763/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
Deprecated==1.2.13
Django==3.2.5
django-js-asset==1.2.2
django-mptt @ file:///home/conda/feedstock_root/build_artifacts/django-mptt_1631114637768/work
django-oauth-toolkit==1.5.0
djangorestframework @ file:///home/conda/feedstock_root/build_artifacts/djangorestframework_1616757768152/work
djangorestframework-simplejwt==4.8.0
drf-yasg==1.20.0
idna @ file:///tmp/build/80754af9/idna_1622654382723/work
importlib-metadata @ file:///C:/ci/importlib-metadata_1617877477759/work
inflection==0.5.1
ipython @ file:///C:/ci/ipython_1617120911285/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
isort @ file:///tmp/build/80754af9/isort_1628603791788/work
itypes @ file:///home/conda/feedstock_root/build_artifacts/itypes_1591807113790/work
jedi @ file:///C:/ci/jedi_1606932640618/work
Jinja2 @ file:///tmp/build/80754af9/jinja2_1624781299557/work
jwcrypto==1.0
lazy-object-proxy @ file:///C:/ci/lazy-object-proxy_1616529288960/work
MarkupSafe @ file:///C:/ci/markupsafe_1621528502553/work
mccabe==0.6.1
mypy-extensions==0.4.3
mysqlclient==2.0.3
oauthlib @ file:///tmp/build/80754af9/oauthlib_1623060228408/work
olefile @ file:///Users/ktietz/demo/mc3/conda-bld/olefile_1629805411829/work
packaging @ file:///tmp/build/80754af9/packaging_1625611678980/work
parso==0.7.0
pathspec==0.7.0
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow @ file:///C:/ci/pillow_1625663294916/work
promise @ file:///C:/ci/promise_1614011696315/work
prompt-toolkit @ file:///Users/ktietz/demo/mc3/conda-bld/prompt-toolkit_1629146841569/work
psycopg2 @ file:///D:/bld/psycopg2_1614021220096/work
pycodestyle @ file:///tmp/build/80754af9/pycodestyle_1615748559966/work
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments @ file:///tmp/build/80754af9/pygments_1629234116488/work
PyJWT @ file:///C:/ci/pyjwt_1619682721924/work
pylint @ file:///C:/ci/pylint_1617136059240/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
PySocks @ file:///C:/ci/pysocks_1605307512533/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-dotenv @ file:///home/conda/feedstock_root/build_artifacts/python-dotenv_1624213301688/work
python3-openid @ file:///tmp/build/80754af9/python3-openid_1614366121783/work
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
PyYAML==5.4.1
regex @ file:///C:/ci/regex_1628063427816/work
requests @ file:///tmp/build/80754af9/requests_1629994808627/work
six @ file:///tmp/build/80754af9/six_1623709665295/work


